+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented here. The format follows
 [Keep a Changelog](https://keepachangelog.com/en/1.1.0/) and
 [Semantic Versioning](https://semver.org/).
 
-### Unreleased
+### v0.8.0 (07-Aug-2026)
 
 #### New features